дз4_18/08

+++ servise/index.js
@@ -1,23 +1,69 @@
 const Contacts = require("./schemas/contact");
+const Users = require("./schemas/users");
 
-const getAllContacts = async () => {
-  return Contacts.find();
+const getAllContacts = async ({ owner, skip, limit, favorite }) => {
+  return Contacts.find(
+    favorite ? { owner, favorite } : { owner },
+    "-createdAt -updatedAt"
+  )
+    .populate("owner", "email subscription -_id")
+    .skip(skip)
+    .limit(limit);
 };
-const getByIdContact = async (id) => {
-  return Contacts.findOne({ _id: id });
+
+const getByIdContact = async ({ id, owner }) => {
+  return Contacts.findOne({ _id: id, owner });
 };
-const createContact = async (body) => {
-  return Contacts.create({ ...body });
+
+const createContact = async ({ body, owner }) => {
+  return Contacts.create({ ...body, owner });
 };
-const removeContact = async (id) => {
-  return Contacts.findByIdAndRemove({ _id: id });
+
+const removeContact = async ({ id, owner }) => {
+  return Contacts.findByIdAndRemove({ _id: id, owner });
 };
-const updateContact = async (id, body) => {
-  return Contacts.findByIdAndUpdate({ _id: id }, body);
+
+const updateContact = async ({ id, body, owner }) => {
+  return Contacts.findByIdAndUpdate({ _id: id, owner }, body, { new: true });
 };
-const updateStatusContact = async (id, body) => {
-  return Contacts.findByIdAndUpdate({ _id: id }, body);
+const updateStatusContact = async ({ id, body, owner }) => {
+  return Contacts.findByIdAndUpdate({ _id: id, owner }, body, { new: true });
 };
+// Домашнє завдання №4  -validateEmail-createUser-updateUserToken-findByIdUser-updateSubscription
+
+const validateEmail = async (email) => {
+  const user = await Users.findOne({ email });
+  return user;
+};
+
+const createUser = async ({ email, password }) => {
+  const result = await Users.create({ email, password });
+  return result;
+};
+
+const updateUserToken = async ({ id, token }) => {
+  const result = await Users.findByIdAndUpdate(
+    id,
+    { token: token },
+    { new: true }
+  );
+  return result;
+};
+
+const findByIdUser = async ({ id }) => {
+  const result = await Users.findById(id);
+  return result;
+};
+
+const updateSubscription = async ({ id, subscription }) => {
+  const result = await Users.findByIdAndUpdate(
+    id,
+    { subscription },
+    { new: true }
+  );
+  return result;
+};
+
 module.exports = {
   getAllContacts,
   getByIdContact,
@@ -25,4 +71,9 @@ module.exports = {
   removeContact,
   updateContact,
   updateStatusContact,
+  validateEmail,
+  createUser,
+  updateUserToken,
+  findByIdUser,
+  updateSubscription,
 };
